Split apart platform-specific dependencies
